% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBAnalysis-package.R
\docType{package}
\name{MBAnalysis-package}
\alias{MBAnalysis}
\alias{MBAnalysis-package}
\title{MBAnalysis: Multiblock Exploratory and Predictive Data Analysis}
\description{
Exploratory and predictive methods for the analysis of several blocks of variables measured on the same individuals.
}
\references{
Tchandao Mangamana, E., Cariou, V., Vigneau, E., Glèlè Kakaï, R. L., & Qannari, E. M. (2019). Unsupervised multiblock data analysis: A unified approach and extensions. Chemometrics and Intelligent Laboratory Systems, 194, 103856.

Tchandao Mangamana, E., Glèlè Kakaï, R., & Qannari, E. M. (2021). A general strategy for setting up supervised methods of multiblock data analysis. Chemometrics and Intelligent Laboratory Systems, 217, 104388.
}
\author{
\strong{Maintainer}: Benjamin Mahieu \email{benjamin.mahieu@oniris-nantes.fr}

Authors:
\itemize{
  \item Essomanda Tchandao Mangamana \email{tchanesso@yahoo.fr}
  \item Evelyne Vigneau \email{evelyne.vigneau@oniris-nantes.fr}
  \item Veronique Cariou \email{veronique.cariou@oniris-nantes.fr}
}

}
\keyword{internal}
