% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.MBPCA.R
\name{print.MBPCA}
\alias{print.MBPCA}
\title{Print of \code{\link{MBPCA}} Objects}
\usage{
\method{print}{MBPCA}(x, ...)
}
\arguments{
\item{x}{An object resulting from \code{\link{MBPCA}}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print of \code{\link{MBPCA}} Objects
}
\examples{
data(ham)
X=ham$X
block=ham$block
res.mbpca <- MBPCA(X,block, name.block=names(block))
print(res.mbpca)

}
\seealso{
\code{\link{summary.MBPCA}}
}
