% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.MBPCA.R
\name{summary.MBPCA}
\alias{summary.MBPCA}
\title{Summary of \code{\link{MBPCA}} Objects}
\usage{
\method{summary}{MBPCA}(object, ...)
}
\arguments{
\item{object}{An object resulting from \code{\link{MBPCA}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The summary.
}
\description{
Edits the Cumulative Explained Variance, Block Explained Variance per Dimension and Block Saliences per Dimension of a \code{\link{MBPCA}} object.
}
\examples{
data(ham)
X=ham$X
block=ham$block
res.mbpca <- MBPCA(X,block, name.block=names(block))
summary(res.mbpca)

}
\seealso{
\code{\link{plot.MBPCA}}
}
