% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.MBWCov.R
\name{summary.MBWCov}
\alias{summary.MBWCov}
\title{Summary of \code{\link{MBWCov}} Objects}
\usage{
\method{summary}{MBWCov}(object, ...)
}
\arguments{
\item{object}{An object resulting from \code{\link{MBWCov}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The summary.
}
\description{
Edits the Cumulative Explained Variance, Block Explained Variance per Dimension and Block Saliences per Dimension of a \code{\link{MBWCov}} object.
}
\examples{
data(ham)
X=ham$X
block=ham$block
Y=ham$Y
res.mbwcov <- MBWCov(X, Y, block, name.block = names(block))
summary(res.mbwcov)

}
\seealso{
\code{\link{plot.MBWCov}}
}
