% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ComDim.R
\name{plot.ComDim}
\alias{plot.ComDim}
\title{Default Plots for \code{\link{ComDim}}}
\usage{
\method{plot}{ComDim}(x, axes = c(1, 2), size = 2.25, ...)
}
\arguments{
\item{x}{An object resulting from \code{\link{ComDim}}.}

\item{axes}{Which dimensions should be plotted?}

\item{size}{The overall size of labels, points, etc.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The default plots.
}
\description{
Successively performs \code{\link{MBplotScores}}, \code{\link{MBplotVars}} and \code{\link{MBplotBlocks}} with the default values of parameters but axes and size.
}
\examples{
data(ham)
X=ham$X
block=ham$block
res.comdim <- ComDim(X,block,name.block=names(block))
plot(res.comdim)

}
\seealso{
\code{\link{MBplotScores}} \code{\link{MBplotVars}} \code{\link{MBplotBlocks}}
}
