% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ComDim.R
\name{print.ComDim}
\alias{print.ComDim}
\title{Print of \code{\link{ComDim}} objects}
\usage{
\method{print}{ComDim}(x, ...)
}
\arguments{
\item{x}{An object resulting from \code{\link{ComDim}}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print of \code{\link{ComDim}} objects
}
\examples{
data(ham)
X=ham$X
block=ham$block
res.comdim <- ComDim(X,block,name.block=names(block))
print(res.comdim)

}
\seealso{
\code{\link{summary.ComDim}}
}
