% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.MBPLS.R
\name{print.MBPLS}
\alias{print.MBPLS}
\title{Print of \code{\link{MBPLS}} objects}
\usage{
\method{print}{MBPLS}(x, ...)
}
\arguments{
\item{x}{An object resulting from \code{\link{MBPLS}}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print of \code{\link{MBPLS}} objects
}
\examples{
data(ham)
X=ham$X
block=ham$block
Y=ham$Y
res.mbpls <- MBPLS(X, Y, block, name.block = names(block))
print(res.mbpls)

}
\seealso{
\code{\link{summary.MBPLS}}
}
