% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.MBWCov.R
\name{print.MBWCov}
\alias{print.MBWCov}
\title{Print of \code{\link{MBWCov}} Objects}
\usage{
\method{print}{MBWCov}(x, ...)
}
\arguments{
\item{x}{An object resulting from \code{\link{MBWCov}}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print of \code{\link{MBWCov}} Objects
}
\examples{
data(ham)
X=ham$X
block=ham$block
Y=ham$Y
res.mbwcov <- MBWCov(X, Y, block, name.block = names(block))
print(res.mbwcov)

}
\seealso{
\code{\link{summary.MBWCov}}
}
