% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ComDim.R
\name{summary.ComDim}
\alias{summary.ComDim}
\title{Summary of \code{\link{ComDim}} Objects}
\usage{
\method{summary}{ComDim}(object, ...)
}
\arguments{
\item{object}{An object resulting from \code{\link{ComDim}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The summary.
}
\description{
Edits the Cumulative Explained Variance, Block Explained Variance per Dimension and Block Saliences per Dimension of a \code{\link{ComDim}} object.
}
\examples{
data(ham)
X=ham$X
block=ham$block
res.comdim <- ComDim(X,block,name.block=names(block))
summary(res.comdim)

}
\seealso{
\code{\link{plot.ComDim}}
}
